+++ include/rain.hpp
@@ -4,17 +4,19 @@
 // Versioning.
 #define RAIN_VERSION_MAJOR 6
 #define RAIN_VERSION_MINOR 0
-#define RAIN_VERSION_REVISION 11
+#define RAIN_VERSION_REVISION 12
 
 #include "rain/build.hpp"
 
 #include "rain/algorithm/kmp.hpp"
-#include "rain/thread/thread-pool.hpp"
+#include "rain/string/command-line-parser.hpp"
+#include "rain/string/waterfall-parser.hpp"
 
 #include "rain/gdiplus.hpp"
 #include "rain/memmem.hpp"
 #include "rain/platform.hpp"
 #include "rain/string.hpp"
+#include "rain/thread-pool.hpp"
 #include "rain/time.hpp"
 #include "rain/types.hpp"
 #include "rain/windows.hpp"