+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 6.0.12
+
+Add `std::error_code` handling to `ThreadPool::queueTask`, which speeds up code significantly if surpassing system thread limits. Move `ThreadPool` to `Rain` namespace.
+
+Restructure `CommandLineParser` and `WaterfallParser`, and add `std::vector` parsing.
+
 ## 6.0.11
 
 Refactor organization in `Algorithm`, `String`, and `Thread`. Remove erroneous `ConditionVariable` implementation. Improve `ThreadPool` implementation to handle common exception upon creating thread in a resource-limited environment. Use `std::condition_variable` correctly in `ThreadPool`. Improve `String` namespace with `anyToAny` and efficient whitespace methods.